worked on adding the emulator
