+++ package.json
@@ -9,6 +9,7 @@
         "@emotion/cache": "^11.10.3",
         "@emotion/react": "^11.10.4",
         "@emotion/styled": "^11.10.4",
+        "@firebase/app": "^0.9.0",
         "@mui/icons-material": "^5.10.16",
         "@mui/lab": "^5.0.0-alpha.100",
         "@mui/material": "^5.10.6",
